refactor: 테스트 필드를 생성자 주입으로 변경
