+++ src/main/java/qna/config/JpaConfig.java
@@ -1,9 +1,9 @@
-package qna;
+package qna.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @Configuration
 @EnableJpaAuditing
-public class JpaConfig {
+public class JpaConfig{
 }
