+++ src/test/java/qna/domain/QuestionRepositoryTest.java
@@ -6,13 +6,15 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import java.util.Optional;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 
 @RepositoryTest
 class QuestionRepositoryTest {
 
-    @Autowired
     private QuestionRepository questionRepository;
+
+    public QuestionRepositoryTest(QuestionRepository questionRepository) {
+        this.questionRepository = questionRepository;
+    }
 
     @Nested
     class FindByIdAndDeleted {
